requirements.txt
setuptools==80.9.0
fastapi==0.115.14
uvicorn==0.34.3
email-validator==2.2.0
pydantic-settings==2.10.1
apscheduler==3.11.0
motor==3.7.1
neo4j==5.28.1
python-dotenv==1.1.1
python-dateutil==2.9.0
python-jose==3.5.0
passlib==1.7.4
pyjwt==2.10.1
aiohttp==3.12.13
bcrypt==4.3.0